Actually merge frontend
