+++ Client/package.json
@@ -1,4 +1,3 @@
-// After merging pls fix
 {
   "name": "vrp-frontend",
   "version": "0.1.0",
@@ -20,6 +19,7 @@
     "react-scripts": "3.3.0",
     "reactstrap": "^8.2.0",
     "redux": "^4.0.5",
+    "redux-saga": "^1.1.3",
     "typescript": "^3.7.4"
   },
   "scripts": {
