--- Client/src/components/MapView.css
@@ -1,8 +1,0 @@
-.lefleat-map {
-    height: 650px;
-}
-
-.container {
-    background-color: white;
-    padding: 1em 0.5em;
-}
