--- Client/src/constants/config.jsx
@@ -1,3 +1,0 @@
-export const ATTRIBUTION = '&amp;copy <a href="http://osm.org/copyright">OpenStreetMap</a> contributors';
-export const MAP_URL = "https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png";
-export const WEBSOCKET_URL = "ws://localhost:5000/ws";
