+++ Client/src/index.jsx
@@ -1,18 +1,12 @@
-// AFTER MERGING PLS FIX
 import React from 'react'
 import { render } from 'react-dom'
 import { Provider } from 'react-redux'
 import { createStore, applyMiddleware } from 'redux'
-import reduxWebsocket, { connect } from '@giantmachines/redux-websocket';
-import rootReducer from './reducers'
-import App from './components/App'
-import { WEBSOCKET_URL } from './constants/config'
+import { BrowserRouter, Switch, Route } from "react-router-dom";
+
 import 'bootstrap/dist/css/bootstrap.css';
 import './index.css';
 
-
-import { BrowserRouter, Switch, Route } from "react-router-dom";
-import store from "./js/store/index";
 import App from "./js/components/App";
 import Register from "./js/components/Register";
 import { NavMenu } from "./js/components/NavMenu";
@@ -20,15 +14,9 @@ import { NavMenu } from "./js/components/NavMenu";
 import ApiAuthorizationRoutes from "./js/components/api-authorization/ApiAuthorizationRoutes";
 import { ApplicationPaths } from "./js/components/api-authorization/ApiAuthorizationConstants";
 
-import "bootstrap/dist/css/bootstrap.css";
+import rootReducer from './js/reducers'
+import store from "./js/store/index";
 
-const reduxWebsocketMiddleware = reduxWebsocket();
-const store = createStore(
-    rootReducer,
-    applyMiddleware(reduxWebsocketMiddleware)
-);
-store.dispatch(connect(WEBSOCKET_URL));
-//const unsubscribe = store.subscribe(() => console.log(store.getState()))
 
 render(
   <Provider store={store}>
