+++ Client/src/js/components/MapView.css
@@ -1,0 +1,3 @@
+.lefleat-map {
+  height: 650px;
+}
