+++ Client/src/js/constants/action-types.jsx
@@ -8,17 +8,16 @@ import {
   WEBSOCKET_SEND,
 } from '@giantmachines/redux-websocket';
 
-export const WEBSOCKET_PREFIX = 'REDUX_WEBSOCKET_DEMO';
+// export const WEBSOCKET_PREFIX = 'REDUX_WEBSOCKET_DEMO';
 
-export const REDUX_WEBSOCKET_BROKEN = `${WEBSOCKET_PREFIX}::${WEBSOCKET_BROKEN}`;
-export const REDUX_WEBSOCKET_OPEN = `${WEBSOCKET_PREFIX}::${WEBSOCKET_OPEN}`;
-export const REDUX_WEBSOCKET_CLOSED = `${WEBSOCKET_PREFIX}::${WEBSOCKET_CLOSED}`;
-export const REDUX_WEBSOCKET_MESSAGE = `${WEBSOCKET_PREFIX}::${WEBSOCKET_MESSAGE}`;
-export const REDUX_WEBSOCKET_CONNECT = `${WEBSOCKET_PREFIX}::${WEBSOCKET_CONNECT}`;
-export const REDUX_WEBSOCKET_DISCONNECT = `${WEBSOCKET_PREFIX}::${WEBSOCKET_DISCONNECT}`;
-export const REDUX_WEBSOCKET_SEND = `${WEBSOCKET_PREFIX}::${WEBSOCKET_SEND}`;
+// export const REDUX_WEBSOCKET_BROKEN = `${WEBSOCKET_PREFIX}::${WEBSOCKET_BROKEN}`;
+// export const REDUX_WEBSOCKET_OPEN = `${WEBSOCKET_PREFIX}::${WEBSOCKET_OPEN}`;
+// export const REDUX_WEBSOCKET_CLOSED = `${WEBSOCKET_PREFIX}::${WEBSOCKET_CLOSED}`;
+// export const REDUX_WEBSOCKET_MESSAGE = `${WEBSOCKET_PREFIX}::${WEBSOCKET_MESSAGE}`;
+// export const REDUX_WEBSOCKET_CONNECT = `${WEBSOCKET_PREFIX}::${WEBSOCKET_CONNECT}`;
+// export const REDUX_WEBSOCKET_DISCONNECT = `${WEBSOCKET_PREFIX}::${WEBSOCKET_DISCONNECT}`;
+// export const REDUX_WEBSOCKET_SEND = `${WEBSOCKET_PREFIX}::${WEBSOCKET_SEND}`;
 
-export const ADD_ARTICLE = "ADD_ARTICLE";
 export const DATA_LOADED = "DATA_LOADED";
 
 export const REGISTER_USER = "REGISTER_USER";
@@ -26,3 +25,7 @@ export const REGISTER_USER = "REGISTER_USER";
 export const DATA_REQUESTED = "DATA_REQUESTED";
 export const API_ERRORED = "API_ERRORED";
 
+export const ADD_PACKAGE = "ADD_PACKAGE";
+export const DELETE_PACKAGE = "DELETE_PACKAGE";
+export const SELECT_PACKAGE = "SELECT_PACKAGE";
+export const CENTER_MAP = "CENTER_MAP";
