+++ Client/src/js/constants/config.jsx
@@ -1,0 +1,3 @@
+export const ATTRIBUTION = '&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors';
+export const MAP_URL = "https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png";
+export const WEBSOCKET_URL = "wss://localhost:5001/ws";
