+++ Client/src/js/index.jsx
@@ -1,5 +1,3 @@
 import store from "../js/store/index";
-import { addArticle } from "../js/actions/index";
 
 window.store = store;
-window.addArticle = addArticle;
