+++ Client/src/js/store/index.jsx
@@ -1,31 +1,27 @@
 import { createStore, applyMiddleware, compose } from "redux";
 import rootReducer from "../reducers/index";
-import { forbiddenWordsMiddleware } from "../middleware";
 import createSagaMiddleware from "redux-saga";
 import apiSaga from "../sagas/api-saga";
-import reduxWebsocket from '@giantmachines/redux-websocket';
-
-import { WEBSOCKET_PREFIX } from "../constants/action-types";
+import reduxWebsocket, { connect } from '@giantmachines/redux-websocket';
+import { WEBSOCKET_URL } from '../constants/config'
 
 const initialiseSagaMiddleware = createSagaMiddleware();
 
 const storeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 
 // Create the middleware instance.
-const reduxWebsocketMiddleware = reduxWebsocket({
-	prefix: WEBSOCKET_PREFIX,
-	onOpen: (socket) => {
-		window.__socket = socket;
-	}
-});
+const reduxWebsocketMiddleware = reduxWebsocket();
 
 const store = createStore(
   rootReducer,
   storeEnhancers(
-		applyMiddleware(forbiddenWordsMiddleware, initialiseSagaMiddleware, reduxWebsocketMiddleware)
-	),
+    applyMiddleware(initialiseSagaMiddleware, reduxWebsocketMiddleware)
+  ),
 );
 
 initialiseSagaMiddleware.run(apiSaga);
+
+store.dispatch(connect(WEBSOCKET_URL));
+// const unsubscribe = store.subscribe(() => console.log(store.getState()))
 
 export default store;
